pyproject.toml
[project]
    dependencies   =[
    "aiosqlite>=0.21.0",
    "fastapi[all]>=0.121.3",
    "loguru-config>=0.1.0",
    "numpy>=2.3.5",
    "pillow>=10.4.0",
    "pydantic-settings>=2.12.0",
    "pyodbc>=5.3.0",
    "sqlalchemy>=2.0.44",
    "surya-ocr>=0.17.0",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
]
    description    ="A tools for otomax software to bulk insert voucher codes into otomax database."
    name           ="mkit-input-voucher"
    readme         ="README.md"
    requires-python=">=3.13"
    version        ="0.1.0"


[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
    "pytest>=8.4.0",
    "pytest-cov>=6.2.1",
    "pytest-html>=4.1.1",
    "pytest-mock>=3.14.1",
    "ruff>=0.11.13",
    "schemathesis>=4.0.15",
    "pytest-asyncio>=1.1.0",
    "pytest-benchmark>=5.2.3",
    "pytest-xdist[psutil]>=3.8.0",
    "pytest-sugar>=1.1.1",
    "pytest-loguru>=0.4.0",
]

[tool.ruff]
builtins = ["_"]
fix = true
show-fixes = true
src = ["src", "tests"]
target-version = "py313"
line-length = 88
lint.fixable = ["ALL"]
lint.unfixable = []
lint.select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "D",   # pydocstyle (docstrings)
    "ANN", # flake8-annotations (type hints)
    "UP",  # pyupgrade (modern Python syntax)
    "B",   # flake8-bugbear (common bugs)
    "SIM", # flake8-simplify (simplify code)
    "C90", # flake8-comprehensions (comprehension optimizations)
    "T20", # flake8-type-checking (type checking)
    "TRY", # flake8-try-except (try-except optimizations)
    "RUF",  # ruff (general code quality)
    "PERF", # ruff performance checks
    "ARG",  # ruff argument checks
    "PLC",  # ruff plugin checks
    "N",   # ruff naming conventions
]
lint.ignore = [
    "D415",    # docstring first line
    "E501",    # panjang baris
    "D100",    # docstring modul
    "D104",    # docstring package
    "D107",    # docstring __init__
    "D101",    # docstring class
    "D102",    # docstring method
    "D105",    # docstring magic method
    "ANN201",  # `cls`
    "ANN202",  # `self`
    "ANN401",  # Any
    "ANN204",  # __init__ return
    "ANN002",  # *args
    "ANN003",  # **kwargs
    "B008", # unused variable
    "TRY003",  # try-except optimizations
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "ANN"]
"scripts/*" = ["D100"]
"__init__.py" = ["D104"]
"_version.py" = ["D", "ANN", "E", "W", "F", "I", "UP", "B"]
"alembic/*.py" = ["ALL"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true
docstring-code-line-length = 60
skip-magic-trailing-comma = false



[tool.pyright]
    exclude=[
        "${python.envPath}/Lib/",
        "**/*.egg-info",
        "**/*.pyi",
        "**/.env",
        "**/.venv",
        "**/.vscode",
        "**/Lib/",
        "**/__pycache__",
        "**/alembic/*",
        "**/dist",
        "**/lib/",
        "**/migrations",
        "**/node_modules",
        "**/tests",
        "**/typings",
        "**/venv",
        ".*",
        "__*",
        "workspaceFolder/Lib/",
    ]
    ignore=["**/typings"]
    include=["src", "tests"]

    pythonPlatform                    ="Windows"
    pythonVersion                     ="3.13"
    reportMissingImports              ="error"
    reportMissingModuleSource         ="none"
    reportUnnecessaryTypeIgnoreComment="information"
    stubPath                          ="typings"
    typeCheckingMode                  ="standard"
    typeshedPath                      ="typings"

[tool.coverage.run]
source = ["src"]
data_file = ".reports/coverage/.coverage"
omit = [
    "*/tests/*",
    "*/_version.py",
    "*/scripts/*",
    "*/__init__.py",
    "*/site-packages/*",
    "*/.venv/*",
    "*/shibokensupport/*",
    "*/pyscript/*",
    "*/signature_bootstrap.py",
    "src/__init__.py",
    "src/_version.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
show_missing = true
skip_covered = false
skip_empty = true

[tool.coverage.html]
directory = ".reports/coverage/html"
skip_covered = false
skip_empty = true

[tool.pytest.ini_options]
    addopts=[
        "--cov-report=html:.reports/coverage",    # generate HTML coverage report
        "--cov-report=term-missing:skip-covered", # show missing lines in terminal, skip fully covered files
        "--cov=src",                          # measure coverage for the 'src' package
        "--disable-warnings",                     # disable warnings during test runs
        "--ff",                                   # run the failures first and then the rest of the tests
        "--html=.reports/tests/report.html",      # generate HTML report
        "--lf",                                   # only re-run the failures
        "--no-header",                            # omit the header from the test output
        "--self-contained-html",                  # make the HTML report self-contained
        "--strict-markers",                       # strict marker checking
        "--tb=short",                             # short traceback format
        # "-n=auto",                                # run tests in parallel using all available CPUs
        "-q",  # verbose output
        "-ra", # show extra test summary info
    ]
    asyncio_mode="auto"
    generate_report_on_test=true
    python_classes=["Test*"]
    python_files=["*_test.py", "test_*.py"]
    python_functions=["test_*"]
    pythonpath=["backend"]
    testpaths=["tests"]

    filterwarnings=[
        "ignore::DeprecationWarning",
        "ignore::PendingDeprecationWarning",
    ]
    log_cli=true
    log_cli_format="%(asctime)s [%(levelname)8s] %(name)s: %(message)s"
    log_cli_level="INFO"
    log_file=".reports/tests/pytest.log"
    markers=[
        "api: mark test as API test",
        "database: mark test as database-related",
        "e2e: mark end to end testing",
        "integration: mark test as integration test",
        "performance: mark test as performance test",
        "regression: mark test as regression test",
        "slow: mark test as slow to run",
        "smoke: mark test as smoke test",
        "ui: mark test as UI test",
        "unit: mark test as unit test",
    ]
    minversion="8.0"
    norecursedirs=[
        "*.egg",
        ".*",
        ".git",
        ".pytest_cache",
        ".reports",
        ".venv",
        "__pycache__",
        "build",
        "dist",
        "venv",
    ]
